# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fintech PWA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfer-rate:
    get:
      operationId: getTransferRate
      tags: [transfers]
      parameters:
        - name: fromCountry
          in: query
          required: true
          schema:
            type: string
        - name: toCountry
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Rate calculation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferRate"
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      responses:
        "200":
          description: Transfer history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
...
components:
  schemas:
...
    Transfer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fromCountry:
          type: string
        toCountry:
          type: string
        amount:
          type: number
        fee:
          type: number
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        createdAt:
          type: string
      required: [id, userId, fromCountry, toCountry, amount, fee, totalAmount, status, createdAt]
    TransferInput:
      type: object
      properties:
        fromCountry:
          type: string
        toCountry:
          type: string
        amount:
          type: number
        recipientName:
          type: string
        recipientAccount:
          type: string
      required: [fromCountry, toCountry, amount]
    TransferRate:
      type: object
      properties:
        fromCountry:
          type: string
        toCountry:
          type: string
        amount:
          type: number
        fee:
          type: number
        feePercent:
          type: number
        totalAmount:
          type: number
        exchangeRate:
          type: number
        receiveCurrency:
          type: string
        receiveAmount:
          type: number
      required: [fromCountry, toCountry, amount, fee, feePercent, totalAmount, exchangeRate, receiveCurrency, receiveAmount]
...